Remove node_modules
Removes the checked-in `node_modules` folder. Having the modules present
in the repo is fine when there are only a few, but as we prepare to add
development libraries, this will add a large amount of library code that
is unneeded when the action is running in production. The intent is to
follow GitHub's recommended workflow using `@vercel/ncc` to build a
clean distribution.