Add workflow test
Adds a "workflow test" to be run via `act`. This is, in essence, a
smoke test to locally verify the action is working as expected. Requires
a version of `act` new enough to support the correct node version
(`node16` at the time of writing).

Run the test via `tests/workflow-test.sh` - this uses a simple workflow
to run the action. The action requires some ENV vars be set as well as
event details being updated in `event.json`. Copy `event.example.json`
to `event.json` and add a real `sha` - this event is passed into the
action as the event payload.

Refer to:

* https://github.com/nektos/act